+++ tests/loops_kernels/loop_block/loop_block_dcmpl-gfx1030.cl
@@ -9,7 +9,7 @@ void loop_general(__global int *data, int block, int x)
     if ((int)block < (int)1) {
         var5 = data + ((get_global_id(0) * block) * 4) / 4;
         var4 = 0;
-        var6 = (ulong)data + (ulong)(((ulong)((ulong)(get_global_id(0) * block) + (ulong)(get_global_id(0) * ((int)block / 2147483648))) + (ulong)(((ulong)get_global_offset(0) + (ulong)0) * block)) * 4);
+        var6 = (ulong)data + (ulong)(((ulong)(get_global_id(0) * block) + (ulong)(get_global_id(0) * ((int)block / 2147483648)) + (ulong)(((ulong)get_global_offset(0) + (ulong)0) * block)) * 4);
         do {
             var0 = *(__global int*)(var5);
             var3 = var0;
@@ -19,6 +19,6 @@ void loop_general(__global int *data, int block, int x)
             *(__global int*)(var5) = var3;
             var5 = var5 + 4 / 4;
             var6 = (ulong)0 + (ulong)(var6);
-        } while (!(block == ((ulong)var4 + (ulong)1)));
+        } while (!((uint)block == (uint)((ulong)var4 + (ulong)1)));
     }
 }
