+++ tests/loops_kernels/simple_loop_kernels/simple_loop_kernels_dcmpl-gfx1010.cl
@@ -10,7 +10,7 @@ void loop_kernel_0(__global uint *data, uint x, uint y, uint unrollingBreaker)
     int var8;
     int var9;
     var8 = 0;
-    if (unrollingBreaker == 0) {
+    if ((uint)unrollingBreaker == (uint)0) {
         do {
             var7 = (ulong)var8 + (ulong)1;
             var14 = var10;
@@ -36,7 +36,7 @@ void loop_kernel_1(__global uint *data, uint x, uint y, uint unrollingBreaker)
     __global uint *var8;
     int var9;
     var5 = unrollingBreaker;
-    if (var5 == 0) {
+    if ((uint)var5 == (uint)0) {
         do {
             var10 = var7;
             var5 = (ulong)var5 + (ulong)(-1);
@@ -46,7 +46,7 @@ void loop_kernel_1(__global uint *data, uint x, uint y, uint unrollingBreaker)
             var8 = var8 + 0 / 4;
             var6 = (var6) * x;
             *(__global uint*)(var10) = var9;
-        } while (!(((ulong)var5 + (ulong)(-1)) == 0));
+        } while (!((uint)((ulong)var5 + (ulong)(-1)) == (uint)0));
     }
 }
 
@@ -60,7 +60,7 @@ void loop_kernel_2(__global uint *data, uint x, uint y, uint unrollingBreaker)
     int var8;
     __global uint *var9;
     var7 = unrollingBreaker;
-    if (var7 == 0) {
+    if ((uint)var7 == (uint)0) {
         var9 = data + (var7 * 4) / 4;
         var10 = data + (0 * 4) / 4;
         do {
@@ -72,6 +72,6 @@ void loop_kernel_2(__global uint *data, uint x, uint y, uint unrollingBreaker)
             var10 = var10 + (-1) / 4;
             var8 = (var8) * x;
             *(__global uint*)(var12) = var11;
-        } while (!(((ulong)var7 + (ulong)(-1)) == 0));
+        } while (!((uint)((ulong)var7 + (ulong)(-1)) == (uint)0));
     }
 }
